requirements.txt
fitparse
moviepy
numpy
pycairo
pytz
pyyaml
scipy
tzlocal
